# openapi/uch-bank-pilot.openapi.yaml
openapi: 3.1.0
info:
  title: Universal Credit Hub Bank Pilot API
  version: 1.0.0-pilot
  description: |
    A controlled integration contract for a bank-owned UCH pilot. It is not a
    production activation or authority to move money, submit a regulatory return,
    make a lending decision, or post accounting entries.
servers:
  - url: https://{bank-approved-host}
    variables:
      bank-approved-host:
        default: sandbox.example-bank.invalid
paths:
...
  /api/npl-cases:
    get:
      summary: List facility-linked NPL cases in the authorised scope
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Case projections with account, borrower, owner and immutable-event summary }
    post:
      summary: Open one controlled case from an existing at-risk credit account
      security: [{ sessionCookie: [] }, { csrfToken: [] }]
      responses:
        '201': { description: Case and server-generated immutable opening event created atomically }
        '409': { description: Case already exists, facility is not at risk, or exposure/currency/country is not reconciled }
  /api/npl-cases/{id}/events:
    get:
      summary: Read the immutable case chronology
      security: [{ sessionCookie: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Newest-first append-only case events }
    post:
      summary: Append an observed case event and update the exposure projection atomically
      security: [{ sessionCookie: [] }, { csrfToken: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Immutable event created; exposure is calculated by the server }
        '400': { description: Invalid/backdated event, missing evidence, non-financial amount or recovery exceeding exposure }
  /api/npl-cases/waterfall/summary:
    get:
      summary: Derive the observed NPL exposure waterfall separately by currency
      security: [{ sessionCookie: [] }]
      parameters:
        - { in: query, name: start, schema: { type: string, format: date } }
        - { in: query, name: end, schema: { type: string, format: date } }
      responses:
        '200': { description: Opening, inflows, cash/legal recoveries, closing, internal difference and authoritative credit-account difference per currency }
        '400': { description: Invalid date range or period over 24 months }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: connect.sid }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }